dp2_requirements.txt
absl-py==2.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
av==14.4.0
black==25.1.0
braceexpand==0.1.7
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@b46f5ac7587d2e1862f8b7b1573179d80dcdd620
cloudpickle==3.1.1
contourpy==1.2.1
cycler==0.12.1
Cython==3.1.2
decorator==4.4.2
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@b15f64ec4429e23a148972175a0207c5a9ab84cf
detectron2-densepose @ git+https://github.com/facebookresearch/detectron2@96c752ce821a3340e27edd51c28a00665dd32a30#subdirectory=projects/DensePose
dominate==2.9.1
-e git+https://github.com/hukkelas/deep_privacy2/@f4d8f09d1eb8f758c89cf1795ae41f20533942d3#egg=dp2
easydict==1.13
einops==0.8.1
einops-exts==0.0.4
eval_type_backport==0.2.2
face_detection @ git+https://github.com/hukkelas/DSFD-Pytorch-Inference@382de8f715f9e4311e5d08241c1725e820cd8dce
facenet-pytorch==2.6.0
fast_pytorch_kmeans==0.2.2
filelock==3.13.1
filterpy==1.4.5
flatbuffers==25.2.10
fonttools==4.58.4
fsspec==2024.6.1
ftfy==6.3.1
fvcore==0.1.5.post20221221
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
hydra-core==1.3.2
idna==3.10
imageio==2.4.1
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iopath==0.1.9
Jinja2==3.1.4
keras==3.10.0
kiwisolver==1.4.7
libclang==18.1.1
lpips==0.1.4
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml_dtypes==0.5.1
motpy @ git+https://github.com/wmuron/motpy@c77f85d27e371c0a298e9a88ca99292d9b9cbe6b
moviepy==0.2.3.5
mpmath==1.3.0
mtcnn==0.1.1
mypy_extensions==1.1.0
namex==0.1.0
networkx==3.2.1
ninja==1.10.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.3.8
portalocker==3.2.0
proglog==0.1.12
protobuf==5.29.5
psutil==7.0.0
pycocotools==2.0.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pyspng==0.1.3
python-dateutil==2.9.0.post0
PyWavelets==1.6.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
resize-right==0.0.2
rich==14.0.0
scikit-image==0.19.3
scipy==1.7.1
sentry-sdk==2.32.0
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tifffile==2024.8.30
timm==0.6.7
tomli==2.2.1
tops @ git+https://github.com/hukkelas/torch_ops.git@f37203980a50a93092e8e4a49de1b318c5005ddd
torch==2.2.2
torch-fidelity==0.3.0
torchaudio==2.5.1+cu121
torchvision==0.17.2
tqdm==4.67.1
triton==2.2.0
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.5.0
validators==0.35.0
wandb==0.20.1
wcwidth==0.2.13
webdataset==0.2.26
Werkzeug==3.1.3
wrapt==1.17.2
yacs==0.1.8
zipp==3.23.0
